# internal/handlers/users.go
package handlers

import (
	"GoProjects/TaskTracker/internal/auth"
	"GoProjects/TaskTracker/internal/models"
	"GoProjects/TaskTracker/internal/store"
	"context"
	"database/sql"
	"encoding/json"
	"errors"
	"github.com/go-chi/chi/v5"
	"github.com/jackc/pgx/v5"
	"net/http"
	"strconv"
)

type UserHandlers struct {
	Store *store.UserStore
}

func RegisterUserRoutes(r chi.Router, s *store.UserStore) {
	h := &UserHandlers{Store: s}

	r.Route("/users", func(r chi.Router) {
		r.Get("/", h.ListUsers)
		r.Post("/", h.CreateUser)
		r.Get("/{id}", h.GetUser)
		r.Put("/{id}", h.UpdateUser)
		r.Delete("/{id}", h.DeleteUser)
	})
}

// ListUsers godoc
// @Summary      Get all users
// @Description  Returns list of all users
// @Tags         users
// @Produce      json
// @Success      200  {array}  models.User
// @Failure      500  {string}  string  "internal error"
// @Router       /users [get]
func (h *UserHandlers) ListUsers(w http.ResponseWriter, r *http.Request) {
	users, err := h.Store.List(context.Background())
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	err = json.NewEncoder(w).Encode(users)
	if err != nil {
		return
	}
}

// GetUser godoc
// @Summary      Get user by ID
// @Description  Returns a single user by their ID
// @Tags         users
// @Produce      json
// @Param        id   path      int  true  "User ID"
// @Success      200  {object}  models.User
// @Failure      400  {string}  string "invalid id"
// @Failure      500  {string}  string "internal error"
// @Router       /users/{id} [get]
func (h *UserHandlers) GetUser(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "invalid id", http.StatusBadRequest)
		return
	}
	user, err := h.Store.Get(context.Background(), id)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			http.Error(w, "user not found", http.StatusNotFound)
			return
		}
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	err = json.NewEncoder(w).Encode(user)
	if err != nil {
		return
	}
}

// CreateUser godoc
// @Summary      Create user
// @Description  Create a new user account
// @Tags         users
// @Accept 		 json
// @Produce      json
// @Param        user   body    models.User  true  "User info"
// @Success      201  {object}  models.User
// @Failure      400  {string}  string "invalid id"
// @Failure      500  {string}  string "internal error"
// @Router       /users [post]
func (h *UserHandlers) CreateUser(w http.ResponseWriter, r *http.Request) {
	var u models.User
	if err := json.NewDecoder(r.Body).Decode(&u); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	hashed, err := auth.HashPassword(u.Password)
	if err != nil {
		http.Error(w, "failed to hashed password", http.StatusInternalServerError)
		return
	}
	u.Password = hashed

	if err := h.Store.Create(context.Background(), &u); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	err = json.NewEncoder(w).Encode(u)
	if err != nil {
		return
	}
}

// DeleteUser godoc
// @Summary      Delete user
// @Description  Delete a user by ID
// @Tags         users
// @Produce      json
// @Param        id   path      int  true  "User ID"
// @Success      204  {string}  string "no content"
// @Failure      400  {string}  string "invalid id"
// @Failure      500  {string}  string "internal error"
// @Router       /users/{id} [delete]
func (h *UserHandlers) DeleteUser(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "invalid id", http.StatusBadRequest)
		return
	}
	err = h.Store.Delete(context.Background(), id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

// UpdateUser godoc
// @Summary      Update user
// @Description  Updates an existing user
// @Tags         users
// @Accept		 json
// @Produce      json
// @Param        id   path      int  true  "User ID"
// @Param        user   body    models.User  true  "User info"
// @Success      200  {object}  models.User
// @Failure      400  {string}  string "invalid id"
// @Failure      404  {string}  string "not found"
// @Failure      500  {string}  string "internal error"
// @Router       /users/{id} [put]
func (h *UserHandlers) UpdateUser(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		http.Error(w, "invalid id", http.StatusBadRequest)
		return
	}
	var u models.User

	if err := json.NewDecoder(r.Body).Decode(&u); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}
	hashed, err := auth.HashPassword(u.Password)
	if err != nil {
		http.Error(w, "failed to hashed password", http.StatusInternalServerError)
		return
	}
	u.Password = hashed

	u.ID = id
	updated, err := h.Store.Update(context.Background(), &u)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			http.Error(w, "user not found", http.StatusNotFound)
			return
		}
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	err = json.NewEncoder(w).Encode(updated)
	if err != nil {
		return
	}
}